refactored package naming and added basic README
