+++ online-store/frontend/pom.xml
@@ -3,7 +3,6 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.github.abdennebi</groupId>
     <artifactId>frontend</artifactId>
     <packaging>jar</packaging>
