--- online-store/frontend/src/main/java/com/github/abdennebi/client/StockClientApi.java
@@ -1,10 +1,0 @@
-package com.github.abdennebi.client;
-
-import feign.Param;
-import feign.RequestLine;
-
-public interface StockClientApi {
-
-    @RequestLine("GET /stocks/{productRef}")
-    Long getProductQuantity(@Param("productRef") String productRef);
-}
