--- online-store/frontend/src/main/java/com/github/abdennebi/config/SwaggerConfig.java
@@ -1,28 +1,0 @@
-package com.github.abdennebi.config;
-
-import com.google.common.base.Predicates;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.service.Contact;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-import java.util.Collections;
-
-@Configuration
-@EnableSwagger2
-public class SwaggerConfig {
-    @Bean
-    public Docket api() {
-        return new Docket(DocumentationType.SWAGGER_2).apiInfo(apiInfo()).select()
-                .apis(Predicates.not(RequestHandlerSelectors.basePackage("org.springframework.boot")))
-                .build();
-    }
-
-    private ApiInfo apiInfo() {
-        return new ApiInfo("Online Store API", "Online Store API", "1.0.0", "", new Contact("", "", ""), "", "", Collections.emptyList());
-    }
-}
