--- online-store/frontend/src/main/java/com/github/abdennebi/domain/ProductWithStock.java
@@ -1,16 +1,0 @@
-package com.github.abdennebi.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-@Builder
-public class ProductWithStock {
-
-    private String productRef;
-    private String productTitle;
-    private String productDescription;
-    private Long quantity;
-}
