--- online-store/frontend/src/main/java/com/github/abdennebi/service/Product.java
@@ -1,18 +1,0 @@
-package com.github.abdennebi.service;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Product {
-
-    private Long id;
-    private String ref;
-    private String title;
-    private String description;
-}
