+++ online-store/frontend/src/main/java/io/cloudnativestack/Frontend.java
@@ -1,0 +1,12 @@
+package io.cloudnativestack;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Frontend {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Frontend.class, args);
+    }
+}
