+++ online-store/frontend/src/main/java/io/cloudnativestack/client/ProductClientApi.java
@@ -1,0 +1,12 @@
+package io.cloudnativestack.client;
+
+import io.cloudnativestack.service.Product;
+import feign.RequestLine;
+
+import java.util.List;
+
+public interface ProductClientApi {
+
+    @RequestLine("GET /products")
+    List<Product> findAllProduct();
+}
