+++ online-store/frontend/src/main/java/io/cloudnativestack/client/StockClientApi.java
@@ -1,0 +1,10 @@
+package io.cloudnativestack.client;
+
+import feign.Param;
+import feign.RequestLine;
+
+public interface StockClientApi {
+
+    @RequestLine("GET /stocks/{productRef}")
+    Long getProductQuantity(@Param("productRef") String productRef);
+}
