+++ online-store/frontend/src/main/java/io/cloudnativestack/domain/ProductWithStock.java
@@ -1,0 +1,16 @@
+package io.cloudnativestack.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+@Builder
+public class ProductWithStock {
+
+    private String productRef;
+    private String productTitle;
+    private String productDescription;
+    private Long quantity;
+}
