+++ online-store/frontend/src/main/java/io/cloudnativestack/service/Product.java
@@ -1,0 +1,18 @@
+package io.cloudnativestack.service;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Product {
+
+    private Long id;
+    private String ref;
+    private String title;
+    private String description;
+}
