+++ online-store/pom.xml
@@ -3,7 +3,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.github.abdennebi</groupId>
+    <groupId>io.cloudnativestack</groupId>
     <artifactId>online-store</artifactId>
     <version>1.0.0-SNAPSHOT</version>
     <packaging>pom</packaging>
