--- online-store/product-service/src/main/java/com/github/abdennebi/domain/Product.java
@@ -1,27 +1,0 @@
-package com.github.abdennebi.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-public class Product {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    private String ref;
-
-    private String title;
-
-    private String description;
-}
