--- online-store/product-service/src/main/java/com/github/abdennebi/domain/ProductRepository.java
@@ -1,13 +1,0 @@
-package com.github.abdennebi.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import javax.transaction.Transactional;
-import java.util.List;
-
-@Transactional
-public interface ProductRepository extends JpaRepository<Product, Long> {
-
-    List<Product> findByRef(String ref);
-
-}
