+++ online-store/product-service/src/main/java/io/cloudnativestack/domain/ProductRepository.java
@@ -1,0 +1,13 @@
+package io.cloudnativestack.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import javax.transaction.Transactional;
+import java.util.List;
+
+@Transactional
+public interface ProductRepository extends JpaRepository<Product, Long> {
+
+    List<Product> findByRef(String ref);
+
+}
