--- online-store/stock-service/src/main/java/com/github/abdennebi/domain/Stock.java
@@ -1,25 +1,0 @@
-package com.github.abdennebi.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Stock {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    private String productRef;
-
-    private Long quantity;
-}
