--- online-store/stock-service/src/main/java/com/github/abdennebi/domain/StockRepository.java
@@ -1,11 +1,0 @@
-package com.github.abdennebi.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import javax.transaction.Transactional;
-
-@Transactional
-public interface StockRepository extends JpaRepository<Stock, Long> {
-
-    Stock findByProductRef(String productRef);
-}
