+++ online-store/stock-service/src/main/java/io/cloudnativestack/domain/Stock.java
@@ -1,0 +1,25 @@
+package io.cloudnativestack.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Stock {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+
+    private String productRef;
+
+    private Long quantity;
+}
