+++ online-store/stock-service/src/main/java/io/cloudnativestack/domain/StockRepository.java
@@ -1,0 +1,11 @@
+package io.cloudnativestack.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import javax.transaction.Transactional;
+
+@Transactional
+public interface StockRepository extends JpaRepository<Stock, Long> {
+
+    Stock findByProductRef(String productRef);
+}
